+++ sdk/test/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @platforma-sdk/test
 
+## 1.22.71
+
+### Patch Changes
+
+- Updated dependencies [121f031]
+- Updated dependencies [e0d8153]
+  - @milaboratories/pl-middle-layer@1.28.16
+
 ## 1.22.70
 
 ### Patch Changes